refactor(presence): consolidate to Tier 1 single mod.rs
